restructured the page structure of connectors
