+++ src/components/EmptyState.tsx
@@ -37,7 +37,7 @@ export function EmptyState({
       } py-12 px-10 flex flex-col items-center justify-center text-center min-h-[250px] ${className}`}
     >
       {/* Icon bubble */}
-      <div className="w-16 h-16 bg-gray-100 rounded-full flex items-center justify-center mb-6">
+      <div className="w-16 h-16 bg-gray-200 rounded-full flex items-center justify-center mb-6">
         <Icon className="w-6 h-6 text-gray-400" />
       </div>
